+++ Projects/Modules/DesignSystem/Sources/Dropdown/JobisDropdownCell.swift
@@ -1,8 +1,25 @@
-//
-//  JobisDropdownCell.swift
-//  DesignSystem
-//
-//  Created by 신주희 on 2/15/26.
-//  Copyright © 2026 com.team.return. All rights reserved.
-//
+import UIKit
 
+class JobisDropdownCell: UITableViewCell {
+
+    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
+        super.init(style: style, reuseIdentifier: reuseIdentifier)
+        setupCell()
+    }
+
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+
+    private func setupCell() {
+        textLabel?.font = .jobisFont(.description)
+        textLabel?.textAlignment = .center
+        selectionStyle = .gray
+        backgroundColor = .white
+    }
+
+    func configure(with text: String, isSelected: Bool) {
+        textLabel?.text = text
+        textLabel?.textColor = isSelected ? .Primary.blue20 : .GrayScale.gray60
+    }
+}
